Partway refactor of button logic
The buttons need to do different things depending on the value of the project textbox, e.g. starting at a different directory, looking for a different file type. Instead of each of these different behaviors being in their own lambda in the entry point, group them together to have 1 function for each row that is passed the project textbox. I guess what happened was I didn't foresee having to do complex logic in each button based on the project textbox; I thought it could all be done in one-liner lambdas for each parameter that needed to be changed. This is a long commit message.